Section lll part b
I added my attempt to represent the environmental decoherence (still in progress though, it is not working as it should)
I also added some derivation on the density matrix representation. I think I should explain a little more the physical meaning behind it and try to understand it better myself as well since I have not worked with it before.