3 qubit code more realistic example
Ive been trying to implement the coherent errors into the 3 qubit code. The difficulty I am facing is the rounding when trying to find the correction state so that I can find the logical state. I am trying to fix the correction, detection, and logical functions to work for this issue. I am currently in the process of working with the correction function because I am not really sure how the output should look since we have some error offset that is added to each logical qubit.