3 qubit code
In the realistic example, instead of applying a inverse of the error to correct, we should just be applying a pauli x operation.
Also added some things in the stabilizer, but its still under work. Im confused on why we need the Zi operation at the end.
Commented out the coherent x correction because i dont think we need it anymore.