Added a generalized reset ancilla function
It should work properly, only issue is that it takes the absolute value of the real and imag parts of the new and old vector. But I think it makes sense since we want to just find what the general state is. Only issue is im not sure about the phases.